requirements.txt
aniso8601==9.0.1
attrs==21.2.0
cffi==1.15.1
click==8.0.3
cryptography==37.0.4
Flask==2.0.2
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
MarkupSafe==2.0.1
marshmallow==3.14.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.0
pytest==7.1.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.26
tomli==2.0.1
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0